fix: BLE状態をReact Contextで画面間共有 (#1, #2)
- BatteryBridgeContext を新規作成し、BLE状態管理とイベントリスナーを
  Context Provider に移動
- useBatteryBridge を Context consumer のシンプルなラッパーに変更
- RootLayout に BatteryBridgeProvider を追加
- connectToDevice で connectionState を 'connecting' に設定 (#2)

Closes #1
Closes #2
