+++ app/_layout.tsx
@@ -4,6 +4,7 @@ import { StatusBar } from 'expo-status-bar';
 import 'react-native-reanimated';
 
 import { useColorScheme } from '@/hooks/use-color-scheme';
+import { BatteryBridgeProvider } from '@/contexts/BatteryBridgeContext';
 
 export const unstable_settings = {
   anchor: '(tabs)',
@@ -14,11 +15,13 @@ export default function RootLayout() {
 
   return (
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
-      <Stack>
-        <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-        <Stack.Screen name="modal" options={{ presentation: 'modal', title: 'Device Detail' }} />
-      </Stack>
-      <StatusBar style="auto" />
+      <BatteryBridgeProvider>
+        <Stack>
+          <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+          <Stack.Screen name="modal" options={{ presentation: 'modal', title: 'Device Detail' }} />
+        </Stack>
+        <StatusBar style="auto" />
+      </BatteryBridgeProvider>
     </ThemeProvider>
   );
 }
